Rename module with aeon algorithms.
