+++ src/_aeon_algorithms/itgr/_impl_extended_component_process.rs
@@ -2,8 +2,8 @@ use biodivine_lib_param_bn::biodivine_std::traits::Set;
 use biodivine_lib_param_bn::symbolic_async_graph::{GraphColoredVertices, SymbolicAsyncGraph};
 use biodivine_lib_param_bn::VariableId;
 
-use crate::aeon::itgr::{BwdProcess, ExtendedComponentProcess, Process, Scheduler};
-use crate::aeon::saturated_reachability::reach_bwd;
+use crate::_aeon_algorithms::itgr::{BwdProcess, ExtendedComponentProcess, Process, Scheduler};
+use crate::_aeon_algorithms::saturated_reachability::reach_bwd;
 
 impl ExtendedComponentProcess {
     pub fn new(
